Add files via upload
Throughout the project, HTML is used to structure the content of each page, ensuring proper headings, paragraphs, lists, and other HTML elements are used appropriately. CSS is utilized to style the website, including typography, colors, layout, and overall visual presentation. Additionally, JavaScript can be employed to add interactivity, such as form validation or dynamic elements, to enhance the user experience.

The website aims to provide an informative and user-friendly experience for visitors who are interested in learning more about the university, exploring its courses, staying updated with the latest news, and easily contacting the university for further inquiries.